requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
markdown==3.7
bleach==6.2.0
